Create `runtime` and `assets` dir to ease install
According to the READMEs, one needs to create those directories to launch the
app. With this commit, those directories are shipped with the repo, so it's
more convenient and less error prone, to launch the app for the first time.

Some additional notes:

* Git doesn't track directories, hence the canonical way to keep an "empty"
directory in a Git repo, is to put an empty hidden `.gitkeep` file in it.
See http://stackoverflow.com/a/7229996/1796345 for more infos.
(Admitedly, I'm making the bet that adding files in the directory won't hinder
 the work of the app)

* Those directories are gitignored. Adding them in the repo don't change that.
  When other files are added in those directories, they still won't appear
  in `git status` & Co

temp